refractoring the parser, removing the http requirement from the parsing and now starting in the ServerBlock class. The NginxConfig class has been revamped to only contain a string containing the complete config file without comments
