+++ config/default.conf
@@ -1,5 +1,4 @@
-http{
-       server{
+server{
                listen 80;
                server_name localhost www.localhost;
 
@@ -33,13 +32,8 @@ http{
                       deny all;
                }
                client_max_body_size 64M;
-       }
-       client_max_body_size 64M;
-       access_log on;
-       error_page stuff; 
-       
-
 }
+
 
 # another example
 # server {
