+++ config/old_nginx_config.hpp
@@ -6,7 +6,7 @@
 /*   By: nhariman <nhariman@student.42.fr>            +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/06/17 15:44:12 by salbregh      #+#    #+#                 */
-/*   Updated: 2022/07/04 18:54:00 by nhariman      ########   odam.nl         */
+/*   Updated: 2022/07/02 16:28:36 by nhariman      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,40 +14,31 @@
 # define NGINX_CONFIG_HPP
 
 #include <iostream>
+#include <map>
+#include <vector>
 #include <fstream>
-#include <string>
 #include "server_block.hpp"
-
-/*
-Coplien form:
-
-class A final
-{
-   public:
-      A ();
-      A (const A &a);
-      ~A ();
-      A & operator = (const A &a);
-};
-*/
 
 class NginxConfig {
 	private:
-		std::string		_config_file;
-
-		NginxConfig(){};
+		struct s_components
+		{
+			bool	error_page;
+			bool	access_log;
+			bool	client_max_body_size;
+			int		server_block;
+		}	components;
+		
+	protected:
+		std::vector<ServerBlock>			server_blocks;
+		int									http_client_max_body_size;
+		std::map<int, std::string>			error_page;
+		bool								access_log;
+		void	ComponentChecker(s_components component);
 
 	public:
-		// coplien form
-		NginxConfig(std::ifstream	configuration_file);
-		NginxConfig(const NginxConfig& obj);
-		~NginxConfig(){};
-		NginxConfig & operator= (const NginxConfig& obj);
-
-		std::string	RemoveComments(std::string str);
-		std::string	GetConfigFile() const;
-		// NginxConfig(); // sets defaults
-		// int		ParseConfigFile(std::ifstream& config_file);
+		NginxConfig(); // sets defaults
+		int		ParseConfigFile(std::ifstream& config_file);
 		// .. bla bla add in conplien
 };
